requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
jinja2==3.1.4
python-multipart==0.0.12
aiofiles==24.1.0
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
